requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
fastapi==0.135.3
fonttools==4.62.1
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
matplotlib==3.10.8
msgpack==1.1.2
numba==0.65.0
numpy==2.2.6
packaging==26.0
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
pyacoustid==1.3.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDejavu==0.2.0
pydub==0.25.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
requests==2.33.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
wavio==0.0.9
